+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.4 (`2023-05-31`)
 - add: nfs-utils for k3s-cluster
+- add: k3s-nfs-provisoner systemd service
 
 # 0.0.3 (`2023-05-25`)
 - upd: networking (host) & k3s-cluster